LayoutTests/storage/indexeddb/modern/opendatabase-versions.html

This test creates a new database with the default version, commits that versionchange transaction, and then reopens it at different versions to make sure the IDBOpenDBRequests behave appropriately.<br>
<script>

if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}

function done()
{
    alert("Done");
    if (window.testRunner)
        testRunner.notifyDone();
}

var request = window.indexedDB.open("VersionTestDatabase");
alert(request + " (firstPhase)");

request.onsuccess = function()
{
    alert("Unexpected success (firstPhase)");
	done();
}
request.onerror = function(e)
{
    alert("Unexpected error (firstPhase)");
	done();
}

request.onupgradeneeded = function(e)
{
    alert("upgradeneeded (firstPhase): old version - " + e.oldVersion + " new version - " + e.newVersion);
    alert(request.transaction);
    request.transaction.oncomplete = function()
    {
        alert("Version change complete (firstPhase). Database version is now - " + request.transaction.db.version);
        request.transaction.db.close();
        secondPhase();
    }
    request.transaction.onabort = function()
    {
        alert("Version change transaction unexpected abort! (firstPhase)");
        done();
    }
    request.transaction.onerror = function()
    {
        alert("Version change transaction unexpected error! (firstPhase)");
        done();
    }
}

function secondPhase()
{
    var request = window.indexedDB.open("VersionTestDatabase", 1);
    alert(request + " (secondPhase)");
    request.onsuccess = function()
    {
        alert("Successfully opened database at version 1 (secondPhase)");
        request.result.close();
        request.result.close(); // Close it twice just for the heck of it
        thirdPhase();
    }
    request.onerror = function(e)
    {
        alert("Unexpected error (secondPhase)" + e);
        done();
    }
    request.onupgradeneeded = function(e)
    {
    	alert("Unexpected upgrade needed (secondPhase)" + e);
    	done();
    }
}

function thirdPhase()
{
    var request = window.indexedDB.open("VersionTestDatabase", 2);
    alert(request + " (thirdPhase)");
    request.onsuccess = function()
    {
        alert("Unexpected success (thirdPhase)");
        done();
    }
    request.onerror = function(e)
    {
        alert("Unexpected error (thirdPhase)");
        done();
    }
    request.onupgradeneeded = function(e)
    {
        alert("upgradeneeded (thirdPhase): old version - " + e.oldVersion + " new version - " + e.newVersion);
        alert(request.transaction);
        request.transaction.oncomplete = function()
        {
            alert("Version change complete (thirdPhase). Database version is now - " + request.transaction.db.version);
            request.transaction.db.close();
            fourthPhase();
        }
        request.transaction.onabort = function()
        {
            alert("Version change transaction unexpected abort! (thirdPhase)");
            done();
        }
        request.transaction.onerror = function()
        {
            alert("Version change transaction unexpected error! (thirdPhase)");
            done();
        }
    } 
}

function fourthPhase()
{
    // We've upgraded to version 2, so version 1 should not be openable.
    var request = window.indexedDB.open("VersionTestDatabase", 1);
    alert(request + " (fourthPhase)");
    request.onsuccess = function()
    {
        alert("Unexpected success (fourthPhase)");
        done();
    }
    request.onerror = function(e)
    {
        alert("Expected error (fourthPhase) - " + request.error.name);
        done();
    }
    request.onupgradeneeded = function(e)
    {
        alert("Unexpected upgradeneeded (fourthPhase)");
        done();
    } 
}

</script>
